+++ back/engines/commercial/mcp_server/app/policies/doorkeeper/application_policy.rb
